+++ tmlt/analytics/query_expr.py
@@ -467,9 +467,7 @@ class JoinPublic(QueryExpr):
                 raise ValueError("Join columns must be distinct")
 
         if isinstance(self.public_table, DataFrame):
-            self.public_table = coerce_spark_schema_or_fail(
-                self.public_table, allow_nan_and_null=True
-            )
+            self.public_table = coerce_spark_schema_or_fail(self.public_table)
 
     def accept(self, visitor: "QueryExprVisitor") -> Any:
         """Visit this QueryExpr with visitor."""